+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `CHANGELOG.md` to track changes and releases
 - _Overview_, _Usage_, _See Also_, and _Roadmap_ sections to `README.md`
 
+### Fixed
+- Prevent _Close Other Tabs_ menu item from being enabled when only pinned tabs are present
+
 ## [1.0.0] - 2020-07-03
 ### Added
 - _Close Other Tabs Item_ extension